Created the endpoint for weather
