+++ src/main/java/files/aiprojectjhams/dto/chatgpt/Message.java
@@ -1,5 +1,4 @@
-
-package files.aiprojectjhams.dto;
+package files.aiprojectjhams.dto.chatgpt;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
@@ -13,37 +12,42 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonPropertyOrder({
-    "cached_tokens",
-    "audio_tokens"
+        "role",
+        "content"
 })
 @Generated("jsonschema2pojo")
-public class PromptTokensDetails {
+public class Message {
 
-    @JsonProperty("cached_tokens")
-    private Integer cachedTokens;
-    @JsonProperty("audio_tokens")
-    private Integer audioTokens;
+    @JsonProperty("role")
+    private String role;
+    @JsonProperty("content")
+    private String content;
     @JsonIgnore
     private Map<String, Object> additionalProperties = new LinkedHashMap<String, Object>();
 
-    @JsonProperty("cached_tokens")
-    public Integer getCachedTokens() {
-        return cachedTokens;
+    public Message(String role, String content) {
+        this.role = role;
+        this.content = content;
     }
 
-    @JsonProperty("cached_tokens")
-    public void setCachedTokens(Integer cachedTokens) {
-        this.cachedTokens = cachedTokens;
+    @JsonProperty("role")
+    public String getRole() {
+        return role;
     }
 
-    @JsonProperty("audio_tokens")
-    public Integer getAudioTokens() {
-        return audioTokens;
+    @JsonProperty("role")
+    public void setRole(String role) {
+        this.role = role;
     }
 
-    @JsonProperty("audio_tokens")
-    public void setAudioTokens(Integer audioTokens) {
-        this.audioTokens = audioTokens;
+    @JsonProperty("content")
+    public String getContent() {
+        return content;
+    }
+
+    @JsonProperty("content")
+    public void setContent(String content) {
+        this.content = content;
     }
 
     @JsonAnyGetter
