+++ src/main/java/files/aiprojectjhams/dto/chatgpt/PromptTokensDetails.java
@@ -1,4 +1,5 @@
-package files.aiprojectjhams.dto;
+
+package files.aiprojectjhams.dto.chatgpt;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
@@ -12,42 +13,37 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonPropertyOrder({
-        "role",
-        "content"
+    "cached_tokens",
+    "audio_tokens"
 })
 @Generated("jsonschema2pojo")
-public class Message {
+public class PromptTokensDetails {
 
-    @JsonProperty("role")
-    private String role;
-    @JsonProperty("content")
-    private String content;
+    @JsonProperty("cached_tokens")
+    private Integer cachedTokens;
+    @JsonProperty("audio_tokens")
+    private Integer audioTokens;
     @JsonIgnore
     private Map<String, Object> additionalProperties = new LinkedHashMap<String, Object>();
 
-    public Message(String role, String content) {
-        this.role = role;
-        this.content = content;
+    @JsonProperty("cached_tokens")
+    public Integer getCachedTokens() {
+        return cachedTokens;
     }
 
-    @JsonProperty("role")
-    public String getRole() {
-        return role;
+    @JsonProperty("cached_tokens")
+    public void setCachedTokens(Integer cachedTokens) {
+        this.cachedTokens = cachedTokens;
     }
 
-    @JsonProperty("role")
-    public void setRole(String role) {
-        this.role = role;
+    @JsonProperty("audio_tokens")
+    public Integer getAudioTokens() {
+        return audioTokens;
     }
 
-    @JsonProperty("content")
-    public String getContent() {
-        return content;
-    }
-
-    @JsonProperty("content")
-    public void setContent(String content) {
-        this.content = content;
+    @JsonProperty("audio_tokens")
+    public void setAudioTokens(Integer audioTokens) {
+        this.audioTokens = audioTokens;
     }
 
     @JsonAnyGetter
