+++ src/main/java/files/aiprojectjhams/dto/chatgpt/Usage.java
@@ -1,5 +1,5 @@
 
-package files.aiprojectjhams.dto;
+package files.aiprojectjhams.dto.chatgpt;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
